employee table has new features, startworkdate, leaveworkdate, isWorking.
Consequently, when an employee leaves the data will be persistent in the
database and the admin can access by checking the /old_employees tab.

This update will impact on the future changes in the project such as
calculating the payroll with specific components which involve the date
that the worker has started so on.